WIP factor out image versions from uploader
