+++ app/uploaders/image_uploader.rb
@@ -1,5 +1,6 @@
 class ImageUploader < WhitehallUploader
   include CarrierWave::MiniMagick
+  include ImageVersions::DefaultImageVersions
 
   configure do |config|
     config.remove_previously_stored_files_after_update = false
@@ -10,34 +11,6 @@ class ImageUploader < WhitehallUploader
     %w[jpg jpeg gif png svg]
   end
 
-  version :s960, if: :use_960_versions? do
-    process resize_to_fill: [960, 640]
-  end
-  version :s712, from_version: :s960, if: :use_960_versions? do
-    process resize_to_fill: [712, 480]
-  end
-  version :s630, from_version: :s960, if: :use_960_versions? do
-    process resize_to_fill: [630, 420]
-  end
-  version :s465, from_version: :s960, if: :use_960_versions? do
-    process resize_to_fill: [465, 310]
-  end
-  version :s300, from_version: :s960, if: :use_960_versions? do
-    process resize_to_fill: [300, 195]
-  end
-  version :s216, from_version: :s960, if: :use_960_versions? do
-    process resize_to_fill: [216, 140]
-  end
-
-  def use_960_versions?(new_file)
-    model.valid_width == 960 && model.valid_height == 640 && bitmap?(new_file)
-  end
-
-  def bitmap?(new_file)
-    return if new_file.nil?
-
-    new_file.content_type !~ /svg/
-  end
 
   def image_cache
     if send("cache_id").present?
